Create 1529A - Eshag Loves Big Arrays.cpp
Output Status : 

By Rajakunalpandit, contest: Codeforces Round #722 (Div. 2), problem: (A) Eshag Loves Big Arrays, Accepted

Eshag has an array a consisting of n integers.

Eshag can perform the following operation any number of times: choose some subsequence of a and delete every element from it which is strictly larger than AVG, where AVG is the average of the numbers in the chosen subsequence.

For example, if a=[1,4,3,2,4] and Eshag applies the operation to the subsequence containing a1, a2, a4 and a5, then he will delete those of these 4 elements which are larger than a1+a2+a4+a54=114, so after the operation, the array a will become a=[1,3,2].

Your task is to find the maximum number of elements Eshag can delete from the array a by applying the operation described above some number (maybe, zero) times.

A sequence b is a subsequence of an array c if b can be obtained from c by deletion of several (possibly, zero or all) elements.

Input
The first line contains an integer t (1≤t≤100) — the number of test cases. The description of the test cases follows.

The first line of each test case contains an integer n (1≤n≤100) — the length of the array a.

The second line of each test case contains n integers a1,a2,…,an (1≤ai≤100) — the elements of the array a.

Output
For each test case print a single integer — the maximum number of elements Eshag can delete from the array a.

Example
inputCopy
3
6
1 1 1 2 2 3
6
9 9 9 9 9 9
6
6 4 1 1 4 1
outputCopy
3
0
3
Note
Consider the first test case.

Initially a=[1,1,1,2,2,3].

In the first operation, Eshag can choose the subsequence containing a1, a5 and a6, their average is equal to a1+a5+a63=63=2. So a6 will be deleted.

After this a=[1,1,1,2,2].

In the second operation, Eshag can choose the subsequence containing the whole array a, the average of all its elements is equal to 75. So a4 and a5 will be deleted.

After this a=[1,1,1].

In the second test case, Eshag can't delete any element.